3. Project 3 - Level 1 -  Data Cleaning (NewYorkCity AirBnb Bookings).ipynb
As the data analyst assigned to this project, I focused on cleaning the AirBnB bookings dataset to ensure data integrity and reliability, as messy data can lead to unreliable outcomes. The cleaning process involved addressing key challenges such as handling missing values through imputation or informed decision-making, identifying and removing duplicate records to maintain uniqueness, and standardizing formatting and units for consistent analysis. Additionally, I detected and addressed outliers that could skew analysis or model performance. This thorough approach to data cleaning was essential for preparing the dataset for accurate and insightful analysis, demonstrating my skills in this critical area of data analysis.